fix: 수익률 출력 값이 커질때 소숫점 이하 출력 안되는거 수정 -> float을 double

+++ src/main/java/lotto/Game.java
@@ -16,7 +16,7 @@ public class Game extends View {
     public static Integer count;
     public static Lotto winning;
     public static Integer bonusNumber;
-    public static Float rate;
+    public static Double rate;
     public Game(){
     }
